+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- loosen version constraint for `python-telegram-bot` to restore compatibility with python3.5
+- fixed return type hints in class `_Persistence`
+- added assertions fixing `mypy` errors
+- tests: fixed compatibility with python3.5-3.7
 
 ## [0.1.0] - 2021-02-14
 ### Added